+++ apps/frontend/src/pages/PageProducts.vue
@@ -1,18 +1,19 @@
 <script>
 import WidgetHeader from '@/components/page-products/WidgetHeader.vue'
-import UiCategoryNavigation from '@/ui/UiCategoryNavigation.vue'
 import VerboseFiltration from '@/components/page-products/VerboseFiltration.vue'
 import WidgetProducts from '@/components/page-products/WidgetProducts.vue'
 import SelectCurrentPage from '@/components/page-products/SelectCurrentPage.vue'
 import WatchedProducts from '@/components/page-products/WatchedProducts.vue'
-import UiLinks from '@/ui/UiLinks.vue'
-import UiTags from '@/ui/UiTags.vue'
-import UiCopyright from '@/ui/UiCopyright.vue'
-import UiTapbar from '@/ui/UiTapbar.vue'
-import UiLinetop from '@/ui/UiLinetop.vue'
 import SelectPriceRanges from '@/components/page-products/SelectPriceRanges.vue'
 import SelectAttributes from '@/components/page-products/SelectAttributes.vue'
 import OffCanvas from '@/components/page-products/OffCanvas.vue'
+
+import UiCategoryNavigation from '@/ui/page-products/UiCategoryNavigation.vue'
+import UiLinks from '@/ui/page-products/UiLinks.vue'
+import UiTags from '@/ui/page-products/UiTags.vue'
+import UiCopyright from '@/ui/page-products/UiCopyright.vue'
+import UiTapbar from '@/ui/page-products/UiTapbar.vue'
+import UiLinetop from '@/ui/page-products/UiLinetop.vue'
 
 import searchProducts from '@/functions/searchProducts'
 import rangeProducts from '@/functions/rangeProducts'
