+++ apps/frontend/src/ui/page-products/UiCategoryNavigation.vue
