+++ apps/frontend/src/ui/page-products/UiLinks.vue
