+++ apps/frontend/src/ui/page-products/UiLogo.vue
