+++ apps/frontend/src/ui/page-products/UiTapbar.vue
